Time: 2 ms (72.42%) | Memory: 19.3 MB (17.56%) - LeetSync

+++ 150-evaluate-reverse-polish-notation/evaluate-reverse-polish-notation.py
@@ -10,8 +10,8 @@ class Solution:
             elif c == "*":
                 stack.append(stack.pop() * stack.pop())
             elif c == "/":
-                a, b = int(stack.pop()), int(stack.pop())
-                stack.append(int(b/a))
+                a, b = stack.pop(), stack.pop()
+                stack.append(int(float(b)/a))
             else:
                 stack.append(int(c))
         return stack[0]
